ANW-28 Agregadas funciones a su propia clase
