+++ Anestesys_BackEnd/controllers/scanMVS.controller.ts
@@ -1,23 +1,58 @@
-import { networkInterfaces } from 'os';
+import { exec } from 'child_process';
+import ping from 'ping';
 
-function detectarDispositivosConectados(): string[] {
-  const interfaces = networkInterfaces();
-  const dispositivosConectados: string[] = [];
+const deviceIP = '192.168.0.26';
 
-  for (const nombreInterface in interfaces) {
-    const interfaz = interfaces[nombreInterface];
-    if (interfaz) {
-      for (const detalle of interfaz) {
-        if (!detalle.internal && detalle.family === 'IPv4') {
-          dispositivosConectados.push(detalle.address);
-        }
+function pingDevice(deviceIP: string): Promise<boolean> {
+  return new Promise<boolean>((resolve, reject) => {
+    ping.sys.probe(deviceIP, (isAlive: boolean | null) => {
+      if (isAlive !== null) {
+        resolve(isAlive);
+      } else {
+        reject(new Error('No se pudo obtener una respuesta del ping.'));
       }
-    }
-  }
-
-  return dispositivosConectados;
+    }, { timeout: 2 });
+  });
 }
 
-// Ejemplo de uso
-const dispositivos = detectarDispositivosConectados();
-console.log('Dispositivos conectados:', dispositivos);
+function getConnectedDevices(callback: (devices: string[]) => void) {
+  exec('arp -a', (error, stdout) => {
+    if (error) {
+      console.error(`Error ejecutando el comando arp: ${error.message}`);
+      return;
+    }
+
+    const lines = stdout.split('\n');
+    const devices: string[] = [];
+
+    lines.forEach(line => {
+      const match = line.match(/(?:[0-9]{1,3}\.){3}[0-9]{1,3}/);
+      
+      if (match) {
+        const deviceIP = match[0];
+        devices.push(deviceIP);
+      }
+    });
+
+    callback(devices);
+  });
+}
+
+pingDevice(deviceIP)
+  .then(isAlive => {
+    if (isAlive) {
+      console.log(`El dispositivo ${deviceIP} está activo.`);
+    } else {
+      console.log(`El dispositivo ${deviceIP} no está activo.`);
+    }
+  })
+  .catch(error => {
+    console.error(`Error al hacer ping al dispositivo ${deviceIP}: ${error.message}`);
+  });
+
+getConnectedDevices(devices => {
+  console.log('Dispositivos conectados:');
+  devices.forEach(device => {
+    console.log(device);
+  });
+});
